fix: issue when drawing player merge tab in data section of settings.
I found that when you have an empty database and you open the settings to Data -> Merge there is an error thrown in the console on mass. This is because of either the variable `comboBox` being `null` or the variable `label` being `null`. So I added a safety check for both. However I would like to note, despite neither of those to variables being reassigned there is no null annotation on the variables and I am unsure as how it was thrown on line 173 and not upon the call of this method.